Qusay template1

+++ JS/skill.js
@@ -40,6 +40,7 @@ tabelElement.addEventListener('click', removeItemFromCart);
 //   };
 // }
 
+let counterConfirm = 0;
 
 function removeItemFromCart(event)
 {
@@ -104,19 +105,22 @@ function addNewSkill(event)
   //   newEducation.render();
   // let data= JSON.stringify(newSkills);
   // localStorage.setItem('Skills',data);
-
+  myForm.reset();
   let confirmWarning = document.getElementById('connfirm');
   let connnfirm = document.getElementById('connnfirm');
   confirmWarning.addEventListener('submit',warningConfirm);
   function warningConfirm()
   {
-
-    let confirmm = confirm('Are You Sure ?');
-    if (confirmm === true) {
-      { let data= JSON.stringify(newSkills);
-        localStorage.setItem('Skills',data);}
-    } else {
-      alert('please reconsidering your inputs');
+    counterConfirm++;
+    if(counterConfirm === 1 )
+    {
+      let confirmm = confirm('Are You Sure ?');
+      if (confirmm === true) {
+        { let data= JSON.stringify(newSkills);
+          localStorage.setItem('Skills',data);}
+      } else {
+        alert('please reconsidering your inputs');
+      }
     }
   }
 }
